Add functions to manage external resources

+++ R/resource.R
@@ -1,18 +1,18 @@
 #' Editor - list External Resources
 #'
-#' List external resources for a study
+#' List external resources for a project
 #'
 #' @return List of external resources
 #' @param idno Project IDNo
 #' @export
-editor_resources_list <- function(idno, api_key=NULL, api_base_url=NULL){
+resources_list <- function(idno, api_key=NULL, api_base_url=NULL){
   
   if(is.null(api_key)){
-    api_key=get_editor_api_key();
+    api_key=get_api_key();
   }
   
   endpoint=paste0('resources/',idno)
-  url=get_editor_api_url(endpoint)
+  url=get_api_url(endpoint)
   
   httpResponse <- GET(url, add_headers("X-API-KEY" = api_key), accept_json(), verbose(get_verbose()))
   output=NULL
@@ -23,11 +23,10 @@ editor_resources_list <- function(idno, api_key=NULL, api_base_url=NULL){
   
   output=list(
     "status_code"=httpResponse$status_code,
-    "response"=nada_http_response_json(httpResponse)
+    "response"=metadataedit_http_response_json(httpResponse)
   )
   
   return (output)
-  
 }
 
 
@@ -38,39 +37,49 @@ editor_resources_list <- function(idno, api_key=NULL, api_base_url=NULL){
 #' @return NULL
 #' @param idno Project IDNo
 #' @param file External resource file to be uploaded
+#' @param resource_type Resource type - documentation, data
 #' @export
-editor_resources_upload <- function(
-    idno,
-    file,
-    api_key=NULL,
-    api_base_url=NULL){
-  
-  endpoint=paste0('files/',idno,'/documentation')
-  
-  if(is.null(api_key)){
-    api_key=get_editor_api_key();
-  }
-  
-  url=get_editor_api_url(endpoint)
-  
-  options=list(
-    "file"=upload_file(file)
-  )
-  
-  httpResponse <- POST(url, add_headers("X-API-KEY" = api_key),body=options, accept_json(), verbose(get_verbose()))
-  output=NULL
-  
-  if(httpResponse$status_code!=200){
-    warning(content(httpResponse, "text"))
-  }
-  
-  output=list(
-    "status_code"=httpResponse$status_code,
-    "response"= nada_http_response_json(httpResponse)
-  )
-  
-  return (output)
-}
+# resources_upload <- function(
+#     idno,
+#     resource_id,
+#     file,
+#     resource_type='documentation',
+#     api_key=NULL,
+#     api_base_url=NULL){
+#   
+#   
+#   valid_resource_types=c("documentation","data")
+#   
+#   if (!resource_type %in% valid_resource_types){
+#     stop(paste("Supported resource types are:",valid_resource_types))
+#   }
+#   
+#   endpoint=paste0('files/',idno,'/', resource_type)
+#   
+#   if(is.null(api_key)){
+#     api_key=get_api_key();
+#   }
+#   
+#   url=get_api_url(endpoint)
+#   
+#   options=list(
+#     "file"=upload_file(file)
+#   )
+#   
+#   httpResponse <- POST(url, add_headers("X-API-KEY" = api_key),body=options, accept_json(), verbose(get_verbose()))
+#   output=NULL
+#   
+#   if(httpResponse$status_code!=200){
+#     warning(content(httpResponse, "text"))
+#   }
+#   
+#   output=list(
+#     "status_code"=httpResponse$status_code,
+#     "response"= metadataedit_http_response_json(httpResponse)
+#   )
+#   
+#   return (output)
+# }
 
 
 
@@ -79,7 +88,7 @@ editor_resources_upload <- function(
 #' Create a new resource
 #'
 #' @return NULL
-#' @param idno \strong{(required)} Study IDNO
+#' @param idno \strong{(required)} Project IDNO
 #' @param dctype Resource document type
 #' @param title Resource title
 #' @param dcformat Resource file format
@@ -100,7 +109,7 @@ editor_resources_upload <- function(
 #'
 #'
 #' @export
-editor_resources_add <- function(
+resources_add <- function(
     idno,
     dctype,
     title,
@@ -118,12 +127,12 @@ editor_resources_add <- function(
     toc=NULL,
     subjects=NULL,
     file_path=NULL,
-    overwrite="no",
+    #overwrite="no",
     api_key=NULL,
     api_base_url=NULL){
   
   if(is.null(api_key)){
-    api_key=get_editor_api_key();
+    api_key=get_api_key();
   }
   
   options=list(
@@ -146,14 +155,14 @@ editor_resources_add <- function(
     overwrite=overwrite
   )
   
-  if (file.exists(file_path)){
+  if (!is.null(file_path) && file.exists(file_path)){
     options$file=upload_file(file_path)
   }
-  else if(is_valid_url(file_path)){
+  else if(!is.null(file_path) && is_valid_url(file_path)){
     options[['filename']]=file_path
   }
   
-  url=get_editor_api_url(paste0('resources/',idno))
+  url=get_api_url(paste0('resources/',idno))
   print(url)
   httpResponse <- POST(url,
                        add_headers("X-API-KEY" = api_key),
@@ -169,7 +178,109 @@ editor_resources_add <- function(
   
   output=list(
     "status_code"=httpResponse$status_code,
-    "response"=nada_http_response_json(httpResponse)
+    "response"=metadataedit_http_response_json(httpResponse)
+  )
+  
+  return (output)
+}
+
+
+#' Editor update resource
+#'
+#' Update resource
+#'
+#' @return NULL
+#' @param idno \strong{(required)} Project IDNO
+#' @param resource_id \strong{(required)} Resource ID
+#' @param dctype Resource document type
+#' @param title Resource title
+#' @param dcformat Resource file format
+#' @param author Author name
+#' @param dcdate Date using YYYY-MM-DD format
+#' @param country Country name
+#' @param language Language or Language code
+#' @param contributor Contributor name
+#' @param publisher Publisher name
+#' @param rights Rights
+#' @param description Resource detailed description
+#' @param abstract  Resource abstract
+#' @param toc Table of contents
+#' @param file_path File path for uploading
+#' @param overwrite Overwrite if resource already exists? Accepted values "yes", "no"
+#'
+#'
+#'
+#'
+#' @export
+resources_update <- function(
+    idno,
+    resource_id,
+    dctype=NULL,
+    title=NULL,
+    subtitle=NULL,
+    dcformat=NULL,
+    author=NULL,
+    dcdate=NULL,
+    country=NULL,
+    language=NULL,
+    contributor=NULL,
+    publisher=NULL,
+    rights=NULL,
+    description=NULL,
+    abstract=NULL,
+    toc=NULL,
+    subjects=NULL,
+    file_path=NULL,
+    api_key=NULL,
+    api_base_url=NULL){
+  
+  if(is.null(api_key)){
+    api_key=get_api_key();
+  }
+  
+  options=list(
+    idno=idno,
+    dctype=dctype,
+    dcformat=dcformat,
+    title=title,
+    subtitle=subtitle,
+    author=author,
+    dcdate=dcdate,
+    country=country,
+    language=language,
+    contributor=contributor,
+    publisher=publisher,
+    rights=rights,
+    description=description,
+    abstract=abstract,
+    toc=toc,
+    subjects=subjects
+  )
+  
+  if (!is.null(file_path) && file.exists(file_path)){
+    options$file=upload_file(file_path)
+  }
+  else if(!is.null(file_path) && is_valid_url(file_path)){
+    options[['filename']]=file_path
+  }
+  
+  url=get_api_url(paste0('resources/',idno, "/", resource_id))
+  print(url)
+  httpResponse <- POST(url,
+                       add_headers("X-API-KEY" = api_key),
+                       body=options,
+                       accept_json(),
+                       verbose(get_verbose()))
+  
+  output=NULL
+  
+  if(httpResponse$status_code!=200){
+    warning(content(httpResponse, "text"))
+  }
+  
+  output=list(
+    "status_code"=httpResponse$status_code,
+    "response"=metadataedit_http_response_json(httpResponse)
   )
   
   return (output)
@@ -185,14 +296,14 @@ editor_resources_add <- function(
 #' @param idno Project IDNo
 #' @param resource_id Resource ID
 #' @export
-editor_resources_delete <- function(idno, resource_id, api_key=NULL, api_base_url=NULL){
+resources_delete <- function(idno, resource_id, api_key=NULL, api_base_url=NULL){
   
   if(is.null(api_key)){
-    api_key=get_editor_api_key();
+    api_key=get_api_key();
   }
   
   endpoint=paste0('resources/delete/',idno,'/',resource_id)
-  url=get_editor_api_url(endpoint)
+  url=get_api_url(endpoint)
   
   httpResponse <- POST(url, add_headers("X-API-KEY" = api_key), accept_json(), verbose(get_verbose()))
   output=NULL
@@ -203,7 +314,7 @@ editor_resources_delete <- function(idno, resource_id, api_key=NULL, api_base_ur
   
   output=list(
     "status_code"=httpResponse$status_code,
-    "response"= nada_http_response_json(httpResponse)
+    "response"= metadataedit_http_response_json(httpResponse)
   )
   
   return (output)
@@ -221,7 +332,7 @@ editor_resources_delete <- function(idno, resource_id, api_key=NULL, api_base_ur
 #' @param skip_uploads TRUE/FALSE - If TRUE, won't upload files
 #' @param overwrite yes/no - Overwrite existing resources
 #' @export
-editor_resources_import_rdf <- function(
+resources_import_rdf <- function(
     idno,
     rdf_file,
     api_key=NULL,
@@ -229,10 +340,10 @@ editor_resources_import_rdf <- function(
 ){
   
   if(is.null(api_key)){
-    api_key=get_editor_api_key();
+    api_key=get_api_key();
   }
   
-  url=get_editor_api_url(paste0('resources/import/',idno))
+  url=get_api_url(paste0('resources/import/',idno))
   
   options=list(
     "file"=upload_file(rdf_file)
@@ -250,7 +361,7 @@ editor_resources_import_rdf <- function(
   
   output=list(
     "status_code"=httpResponse$status_code,
-    "response"= nada_http_response_json(httpResponse)
+    "response"= metadataedit_http_response_json(httpResponse)
   )
   
   return (output)
